Addressed comments on LGConv and LightGCN
