--- torch_geometric/nn/conv/lightgcn_conv.py
@@ -1,63 +1,0 @@
-from torch_geometric.typing import Adj
-
-import torch
-from torch import Tensor
-from torch_geometric.nn import MessagePassing
-
-
-class LightGCNConv(MessagePassing):
-    r"""The neighbor aggregation operator from the `"LightGCN: Simplifying and
-    Powering Graph Convolution Network for Recommendation"
-    <https://arxiv.org/abs/2002.02126#>`_ paper
-
-    .. math::
-        \textbf{e}_u^{(k+1)} =
-        \sum_{i \in N_u} \frac{1}{\sqrt{|N_u|} \sqrt{|N_i|}} \textbf{e}_i^{(k)}
-
-    .. math::
-        \textbf{e}_i^{(k+1)} =
-        \sum_{i \in N_i} \frac{1}{\sqrt{|N_i|} \sqrt{|N_u|}} \textbf{e}_u^{(k)}
-
-    Args:
-        in_channels (int): Size of each input sample, or :obj:`-1` to derive
-            the size from the first input(s) to the forward method.
-        out_channels (int): Size of each output sample.
-        num_users (int): Number of users for recommendation.
-        num_items (int): Number of items to recommend.
-        **kwargs (optional): Additional arguments of
-            :class:`torch_geometric.nn.conv.MessagePassing`.
-    """
-    def __init__(self, in_channels: int, out_channels: int,
-                 num_users: int, num_items: int, **kwargs):
-        super(LightGCNConv, self).__init__(**kwargs)
-
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-
-        self.num_users = num_users
-        self.num_items = num_items
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        pass  # There are no layer parameters to learn.
-
-    def forward(self, x: Tensor, edge_index: Adj) -> Tensor:
-        """Performs neighborhood aggregation for user/item embeddings."""
-        user_item = \
-            torch.zeros(self.num_users, self.num_items, device=x.device)
-        user_item[edge_index[:, 0], edge_index[:, 1]] = 1
-        user_neighbor_counts = torch.sum(user_item, axis=1)
-        item_neightbor_counts = torch.sum(user_item, axis=0)
-        # Compute weight for aggregation: 1 / sqrt(N_u * N_i)
-        weights = user_item / torch.sqrt(
-            user_neighbor_counts.repeat(self.num_items, 1).T
-            * item_neightbor_counts.repeat(self.num_users, 1))
-        weights = torch.nan_to_num(weights, nan=0)
-        out = torch.concat((weights.T @ x[:self.num_users],
-                            weights @ x[self.num_users:]), 0)
-        return out
-
-    def __repr__(self):
-        return '{}({}, {})'.format(self.__class__.__name__, self.in_channels,
-                                   self.out_channels)
